[rust-error] Refactor structure of oneil_error and rename structure to `OneilError`
